+++ src/main/java/com/qcloud/Module/CloudAudit.java
@@ -1,0 +1,12 @@
+package com.qcloud.Module;
+
+/*
+ * see document: <a
+ * href="https://cloud.tencent.com/document/product/629">
+ * https://cloud.tencent.com/document/product/629</a>
+ */
+public class CloudAudit extends Base {
+    public CloudAudit(){
+        serverHost = "cloudaudit.api.qcloud.com";
+    }
+}